\docType{data}
\name{aerolineas}
\alias{aerolineas}
\title{Nombres de aerolineas}
\format{Un data.frame con 16 lineas y 2 variables
\describe{
\item{codigo_carrier}{abreviacion de dos caracteres del nombre de la aerolinea}
\item{nombre}{nombre completo de la aerolinea}
}}
\usage{aerolineas}
\description{Nombres de aerolineas y su respectivo codigo carrier de dos digitos.}
\keyword{datasets}
