\docType{data}
\name{aeropuertos}
\alias{aeropuertos}
\title{Datos de aeropuertos}
\format{Un data.frame con 1,458 lineas y 8 variables
\describe{
\item{codigo_aeropuerto}{codigo FAA del aeropuerto}
\item{nombre}{nombre del aeropuerto}
\item{latitud}{coordenada de ubicacion del aeropuerto}
\item{longitud}{coordenada de ubicacion del aeropuerto}
\item{altura}{altura medida en pies}
\item{zona_horaria}{diferencia horaria respecto del meridiano de Greenwich (GMT)}
\item{horario_verano}{horario legal de acuerdo a las politicas de ahorro de energia. A = Zona Horaria de Ahorro de Luz de Dia (DST) estandar de los EEUU: comienza el segundo Sabado de Marzo y termina el primer Sabado de Noviembre. U = Desconocido. N = No tiene zona horaria especial.}
\item{zona_horaria_iana}{zona horaria IANA, la cual se determina de acuerdo al servicio web GeoNames}
}}
\usage{aeropuertos}
\description{Metadatos utiles de los aeropuertos}
\keyword{datasets}
