\docType{data}
\name{atmosfera}
\alias{atmosfera}
\title{Datos atmosféricos}
\format{Un data.frame con 41472 filas y 11 columnas
\describe{
\item{latitud}{Ubicación de la medida. Grilla espacial de 24 por 24 desde longitud 113.8O a 56.2O y desde latitud 36.2N a 21.2S.}
\item{longitud}{Ubicación de la medida. Grilla espacial de 24 por 24 desde longitud 113.8O a 56.2O y desde latitud 36.2N a 21.2S.}
\item{anio}{72 puntos en el tiempo. Una vez al mes desde enero de 1995 hasta diciembre de 2000.}
\item{mes}{72 puntos en el tiempo. Una vez al mes desde enero de 1995 hasta diciembre de 2000.}
\item{temp_superficie}{Temperatura media mensual basada en la energía emitida desde la Tierra en condiciones de cielo despejado (en K).}
\item{temperatura}{Temperatura media mensual del aire cerca de la superficie de la Tierra (en K).}
\item{presion}{Presión media mensual en una ubicación determinada de la superficie de la Tierra (en mb).}
\item{ozono}{Cantidad media mensual de ozono en la columna atmosférica (en Dobsons).}
\item{nube_baja}{Porcentaje de la media mensual de cielo cubierto por nubes con una presión superior mayor a 680 mb o aproximadamente menor que 3.24 km.}
\item{nube_media}{Porcentaje de la media mensual de cielo cubierto por nubes con una presión superior entre 440-680 mb o aproximadamente menor que 3.24 a 6.5 km.}
\item{nube_alta}{Porcentaje de la media mensual de cielo cubierto por nubes con una presión superior menor o igual a 440 mb o aproximadamente mayor que 6.5 km.}
}}
\usage{atmosfera}
\description{Datos atmosféricos}
\keyword{datasets}
