\docType{data}
\name{aviones}
\alias{aviones}
\title{Datos de aviones}
\format{Un data.frame con 3,322 lineas y 9 variables
\describe{
\item{codigo_cola}{codigo de cola}
\item{anio}{ano de fabricacion}
\item{tipo}{tipo de avion}
\item{fabricante}{fabricante del avion}
\item{modelo}{modelo del avion}
\item{motores}{cantidad de motores}
\item{asientos}{cantidad de asientos}
\item{velocidad}{velocidad crucero promedio medida en millas por hora}
\item{tipo_motor}{tipo de motor del avion}
}}
\usage{aviones}
\description{Metadatos de los aviones y sus codigos de cola en el registro de aviacion de la FAA. American Airways (AA) y Envoy Air (MQ) reportan numero de flota en lugar de numero de cola, por lo que no es posible trazarlos.}
\keyword{datasets}
