\docType{data}
\name{bateadores}
\alias{bateadores}
\title{Tabla de bateadores}
\format{Un data frame con 102.816 filas y 22 columnas
\describe{
\item{ID_jugador}{ID del jugador}
\item{ID_anio}{Año}
\item{orden_equipos}{Orden en que el jugador se movió entre equipos dentro de la misma temporada}
\item{ID_equipo}{ID del equipo (factor)}
\item{ID_liga}{ID de la liga (factor con niveles AA, AL, FL, NL, PL, UA)}
\item{juegos}{Número de juegos en los que participó el jugador}
\item{al_bate}{Turnos al bate}
\item{carreras}{Carreras anotadas}
\item{golpes}{Golpes: cantidad de veces que se alcanzó la base debido a un golpe, sin error de la defensa}
\item{dobles}{Golpes en los que el bateador llegó a segunda base}
\item{triples}{Golpes en los que el bateador llegó a tercera base}
\item{cuadrangulares}{Cuadrangulares - jonrones}
\item{carreras_empujadas}{Carreras empujadas}
\item{bases_robadas}{Bases robadas}
\item{atrapado_robo}{Atrapado en intento de robo}
\item{bases_bolas}{Bases por bolas}
\item{ponches}{Ponches}
\item{bases_intencional}{Bases por bolas intencionales}
\item{golpeado}{Golpeado por lanzamiento}
\item{sacrificios_golpeados}{Sacrificios golpeados}
\item{vuelos_sacrificio}{Vuelos de sacrificio}
\item{doble_matanza}{Bateos para doble matanza}
}}
\description{Estadísticas de bateadores}
\keyword{datasets}
