\docType{data}
\name{comunes}
\alias{comunes}
\title{Modelos comunes}
\format{Un data.frame con 347 filas y 4 columnas
\describe{
\item{fabricante}{Fabricante}
\item{modelo}{Modelo}
\item{modelos}{Número total de modelos}
\item{modelos_anios}{Número total de modelos-años}
}}
\usage{comunes}
\description{Estos modelos tienen al menos 10 años de datos. Este dataset es adecuado para efectuar un join que restringa el número de observaciones a un nivel manejable.}
\keyword{datasets}
