% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datos-package.R
\docType{package}
\name{datos-package}
\alias{datos}
\alias{datos-package}
\title{datos: Traduce al Espaniol Varios Conjuntos de Datos de Practica}
\description{
Provee una version traducida de los siguientes
    conjuntos de datos: diamonds, gapminder, mpg, mtcars, presidential,
    gss-cat, Batting and flights.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cienciadedatos/datos}
  \item Report bugs at \url{https://github.com/cienciadedatos/datos/issues}
}

}
\author{
\strong{Maintainer}: Edgar Ruiz \email{edgararuiz@gmail.com}

Authors:
\itemize{
  \item Riva Quiroga \email{riva.quiroga@uc.cl}
  \item Mauricio Vargas \email{mvargas@dcc.uchile.cl}
  \item Mauro Lepore \email{maurolepore@gmail.com}
}

}
\keyword{internal}
