% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringr.R
\docType{data}
\name{datos-stringr}
\alias{datos-stringr}
\alias{oraciones}
\alias{frutas}
\alias{palabras}
\title{Vectores de caracteres dentro del paquete stringr}
\format{
Un vector

An object of class \code{character} of length 50.

An object of class \code{character} of length 25.

An object of class \code{character} of length 1000.
}
\usage{
oraciones

frutas

palabras
}
\description{
Muestra de vectores de caracteres para practicar la manipulación de cadenas (strings)
}
\details{
oraciones - es una selección y adaptación al español de los datos de práctica contenidos en stringr::sentences.

fruta - es una selección y adaptación al español de los datos de práctica contenidos en stringr::fruit.

palabras - es una selección de las 1000 palabras más frecuentes del Corpus de Referencia del Español Actual (CREA) de la RAE, disponible en github.com/rivaquiroga/frecuencias-crea
}
\examples{
length(oraciones)
oraciones[1:5]

length(frutas)
frutas[1:5]

length(palabras)
palabras[1:5]
}
\keyword{datasets}
