\docType{data}
\name{datos_credito}
\alias{datos_credito}
\title{Datos de crédito}
\format{Un data.frame con 14 filas y 4454 columnas
\describe{
\item{Estado}{estado del crédito}
\item{Antiguedad}{Antigüedad laboral}
\item{Vivienda}{tipo de propiedad de la vivienda}
\item{Plazo}{Plazo del crédito en meses}
\item{Edad}{edad del cliente}
\item{EstadoCivil}{estado civil}
\item{Registros}{existencia de registros previos}
\item{Trabajo}{tipo de trabajo}
\item{Gastos}{cantidad o monto de gastos}
\item{Ingresos}{cantidad o monto de ingresos}
\item{Activos}{cantidad o monto de activos}
\item{Deuda}{cantidad o monto de deudas}
\item{Cantidad}{cantidad solicitada de préstamo}
\item{Precio}{precio del crédito}
}}
\description{Datos asociados a créditos de consumo.}
\keyword{datasets}
