\docType{data}
\name{encuesta}
\alias{encuesta}
\title{Muestra de variables categoricas de una encuesta social}
\format{Un dataframe con 21,473 lineas y 9 variables
\describe{
\item{anio}{ano de la encuesta, 2000-2014}
\item{estado_civil}{estado civil}
\item{edad}{edad. Edad maxima truncada a 89 anos}
\item{raza}{categorias raciales segun la encuesta}
\item{ingreso}{ingreso reportado por los encuestados (US$)}
\item{partido}{partido politico con el que se identifica}
\item{religion}{religion}
\item{denominacion}{denominacion religiosa a la que pertenece}
\item{horas_tv}{horas de television que ve al dia}
}}
\usage{encuesta}
\description{Muestra de variables categoricas de la Encuesta Social General de EE.UU.}
\keyword{datasets}
