\docType{data}
\name{fiel}
\alias{fiel}
\title{Datos del geiser Viejo Fiel (Old Faithful)}
\format{Un data frame con 272 observaciones y 2 variables
\describe{
\item{erupciones}{Tiempo de erupciones, en minutos}
\item{espera}{Minutos entre erupciones}
}}
\usage{fiel}
\description{Datos de tiempo de duración y espera entry erupciones del Viejo Fiel (Old Faithful), es un géiser del Parque Nacional de Yellowstone, situado en Wyoming, Estados Unidos.}
\keyword{datasets}
