\docType{data}
\name{fildeo}
\alias{fildeo}
\title{Tabla de Fildeo}
\format{Un data frame con 140.921 filas y 18 columnas
\describe{
\item{id_jugador}{ID del jugador}
\item{id_anio}{Año}
\item{orden_equipos}{Orden en que el jugador se movió entre equipos dentro de la misma temporada}
\item{id_equipo}{ID del equipo (factor)}
\item{id_liga}{ID de la liga (factor con niveles AA, AL, FL, NL, PL, UA)}
\item{posicion}{Posición del jugdador a la defensiva}
\item{juegos_jugados}{Juegos jugados}
\item{juegos_iniciados}{Juegos iniciados}
\item{InnOuts}{Tiempo jugado por el jugador, expresado en outs}
\item{outs_realizados}{Outs realizados por el jugador}
\item{asistencias}{Asistencias realizadas por el jugador}
\item{errores}{Errores realizados por el jugador}
\item{doble_matanzas}{Doble matanzas en las que el jugador ha estado envuelto defensivamente}
\item{bolas_pasadas}{Bolas dejadas pasar (aplicable a los receptores)}
\item{lanzamientos_desviados}{Lanzamientos desviados (aplicable a los receptores)}
\item{bases_robadas}{Bases robadas por oponente (aplicable a los receptores)}
\item{atrapados_robando}{Oponentes atrapados en intento de robo (aplicable a los receptores)}
\item{clasificacion_zona}{Porcentaje de pelotas atrapadas por un fildeador, en su zona defensiva "típica"}
}}
\description{Estadísticas de jardineros}
\keyword{datasets}
