\docType{data}
\name{flores}
\alias{flores}
\title{Datos sobre la flor Iris de Edgar Anderson}
\format{Un dataframe con 150 casos (filas) y 5 variables (columnas)
\describe{
\item{Largo.Sepalo}{Largo del sépalo}
\item{Ancho.Sepalo}{Ancho del sépalo}
\item{Largo.Petalo}{Largo del pétalo}
\item{Ancho.Petalo}{Ancho del pétalo}
\item{Especies}{A qué especie de la flor Iris corresponde (setosa, versicolor, virginica)}
}}
\usage{flores}
\description{El set de datos sobre la flor Iris entrega las medidas en centímetros del largo y ancho de los sépalos y los pétalos de 50 flores de 3 especies de Iris (setosa, versicolor y virginica).}
\keyword{datasets}
