\docType{data}
\name{jardineros}
\alias{jardineros}
\title{Tabla de jardineros}
\format{Un data frame con 140.921 filas y 18 columnas
\describe{
\item{id_jugador}{ID del jugador}
\item{id_anio}{Año}
\item{orden_equipos}{Orden en que el jugador se movió entre equipos dentro de la misma temporada}
\item{id_equipo}{ID del equipo (factor)}
\item{id_liga}{ID de la liga (factor con niveles AA, AL, FL, NL, PL, UA)}
\item{posicion}{Posición}
\item{juegos}{Número de juegos en los que participó el jugador}
\item{juegos_iniciados}{Juegos iniciados}
\item{InnOuts}{Time played in the field expressed as outs}
\item{PO}{Putouts}
\item{A}{Assists}
\item{E}{Errors}
\item{DP}{Double Plays}
\item{PB}{Passed Balls (by catchers)}
\item{WB}{Wild Pitches (by catchers)}
\item{SB}{Opponent Stolen Bases (by catchers)}
\item{CS}{Opponents Caught Stealing (by catchers)}
\item{ZR}{Zone Rating}
}}
\description{Estadísticas de jardineros}
\keyword{datasets}
