\docType{data}
\name{lanzadores}
\alias{lanzadores}
\title{Tabla de lanzadores}
\format{Un data frame con 46.699 filas y 30 columnas
\describe{
\item{id_jugador}{ID del jugador}
\item{id_anio}{Año}
\item{orden_equipos}{Orden en que el jugador se movió entre equipos dentro de la misma temporada}
\item{id_equipo}{ID del equipo (factor)}
\item{id_liga}{ID de la liga (factor con niveles AA, AL, FL, NL, PL, UA)}
\item{juegos_ganados}{Número de juegos ganados en los que participó el jugador}
\item{juegos_perdidos}{Número de juegos perdidos en los que participó el jugador}
\item{juegos}{Número de juegos en los que participó el jugador}
\item{juegos_iniciados}{Número de juegos en los que ha participado el jugador}
\item{juegos_completados}{Número de juegos completos en los que ha participado el jugador}
\item{juegos_terminados}{Número de juegos en que los pitchers logran que el rival no complete la carrera}
\item{juegos_salvados}{Número de juegos en que el lanzador logra una de las siguientes condiciones independientes: (i) entra al juego con una ventaja no mayor a tres carreras; (ii) entra en el juego, independientemente de la cuenta, con una carrera potencial de empate en base, al bate o en el círculo de espera; (iii) lanza tres entradas.}
\item{entradas_consecutivas}{Número de entradas del lanzador multiplicado por tres.}
\item{golpes}{Golpes: cantidad de veces que se alcanzó la base debido a un golpe, sin error de la defensa}
\item{carreras_ganadas}{Número de carreras ganadas.}
\item{cuadrangulares}{Cuadrangulares - jonrones}
\item{bases_bolas}{Bases por bolas}
\item{ponchados}{Número de bateadores con tres cuentas negativas seguidas (strikes)}
\item{promedio_bateo_rival}{Prometio del total de acierto del rival en los turnos al bate}
\item{promedio_carreras_ganadas}{Promedio de carreras ganadas.}
\item{bases_intencional}{Bases por bolas intencionales}
\item{lanzamientos_descontrolados}{Número de lanzamientos descontrolados}
\item{golpeado}{Golpeado por lanzamiento}
\item{numero_interferencias}{Número de interferencias}
\item{BFP}{Batters faced by Pitcher}
\item{GF}{Games Finished}
\item{R}{Runs Allowed}
\item{SH}{Sacrifices by opposing batters}
\item{SF}{Sacrifice flies by opposing batters}
\item{doble_matanza}{Bateos para doble matanza}
}}
\description{Estadísticas de lanzadores}
\keyword{datasets}
