\docType{data}
\name{millas}
\alias{millas}
\title{Datos de economia de combustible de 1999 y 2008 para 38 modelos populares de automoviles}
\format{Un dataframe con 234 filas y 11 variables
\describe{
\item{fabricante}{fabricante}
\item{modelo}{nombre del modelo}
\item{motor}{tamano del motor del automovil, en litros}
\item{anio}{ano de fabricacion}
\item{cilindros}{numero de cilindros}
\item{transmision}{tipo de transmision}
\item{traccion}{tipo de traccion (d = delantera, t = trasera, 4 = 4 ruedas)}
\item{ciudad}{millas por galon de combustible en ciudad}
\item{autopista}{millas por galon de combustible en autopista}
\item{combustible}{tipo de combustible (p = premium, r = regular, e = etanol, d = diesel, g = gas natural comprimido)}
\item{clase}{tipo de auto}
}}
\usage{millas}
\description{Este conjunto de datos contiene un subconjunto de los datos de economia de combustible que la Agencia de Proteccion Medioambiental (EPA) pone a disposicion en http://fueleconomy.gov. Contiene solo modelos que tuvieron una nueva version cada ano entre 1999 y 2008, lo que fue utilizado como un proxy de la popularidad del automovil.}
\keyword{datasets}
