\docType{data}
\name{mtautos}
\alias{mtautos}
\title{Pruebas de ruta de automoviles de Motor Trend}
\format{Un data frame con 32 observaciones y 11 variables
\describe{
\item{millas}{millas por galon de Estados Unidos}
\item{cilindros}{numero de cilindros}
\item{cilindrada}{suma del volumen util de todos los cilindros del motor en pulgadas cubicas}
\item{caballos}{caballos de fuerza brutos}
\item{eje}{relacion del eje de transmision trasero}
\item{peso}{peso (1000 libras)}
\item{velocidad}{tiempo en recorrer 1/4 de milla}
\item{forma}{forma del motor (en V o en linea)}
\item{transmision}{tipo de transmision (0 = automatico, 1 = manual)}
\item{cambios}{numero de cambios de la caja de cambios}
\item{carburadores}{numero de carburadores}
}}
\usage{mtautos}
\description{Los datos fueron extraidos de la revista Motor Trend de Estados Unidos de 1974, y tiene datos de consumo de combustible y 10 aspectos de diseno y performance de automoviles para 32 automoviles (modelos de 1973-73).}
\keyword{datasets}
