\docType{data}
\name{nombres}
\alias{nombres}
\title{Nombres de bebés}
\format{Un data.frame con 1924665 observaciones y 5 variables
\describe{
\item{anio}{Año del registro}
\item{sexo}{Sexo del bebé}
\item{nombre}{Nombre}
\item{FALSE}{Cantidad de ocurrencias}
\item{prop}{Proporción}
}}
\usage{nombres}
\description{Datos completos sobre nombres de bebés provistos por la Social Security Administration de EE.UU. Incluye todos los nombres con al menos 5 usos.}
\keyword{datasets}
