\docType{data}
\name{oraciones}
\alias{oraciones}
\title{Muestra de vector de caracteres para practicar manipulacion de cadenas de caracteres.}
\format{Una cadena de caracteres con 720 elementos.
\describe{
}}
\usage{oraciones}
\description{oraciones es una coleccion de "Harvard sentences" usada para pruebas estandarizadas de voz.}
\keyword{datasets}
