\docType{data}
\name{paises}
\alias{paises}
\title{Datos de Gapminder}
\format{El dataframe principal tiene 1704 filas y 6 variables
\describe{
\item{pais}{paises incluidos (factor con 142 niveles)}
\item{continente}{continentes (factor con 5 niveles)}
\item{anio}{desde 1952 a 2007, datos cada 5 anos}
\item{esperanza_de_vida}{esperanza de vida al nacer, en anos}
\item{poblacion}{poblacion}
\item{pib_per_capita}{PIB per capita (en dolares US, ajustados segun inflacion)}
}}
\usage{paises}
\description{Extracto de datos de Gapminder sobre expectativa de vida, PIB per capita y poblacion, segun pais}
\source{http://www.gapminder.org/data/}
\keyword{datasets}
