\docType{data}
\name{personas}
\alias{personas}
\title{Tabla de personas}
\format{Un data frame con 19.617 filas y 26 columnas
\describe{
\item{id_jugador}{ID del jugador}
\item{anio_nacimiento}{Año de nacimiento}
\item{mes_nacimiento}{Mes de nacimiento}
\item{dia_nacimiento}{Día de nacimiento}
\item{pais_nacimiento}{País de nacimiento}
\item{estado_nacimiento}{Estado de nacimiento}
\item{ciudad_nacimiento}{Ciudad de nacimiento}
\item{anio_fallecimiento}{Año de fallecimiento}
\item{mes_fallecimiento}{Mes de fallecimiento}
\item{dia_fallecimiento}{Día de fallecimiento}
\item{pais_fallecimiento}{País de fallecimiento}
\item{estado_fallecimiento}{Estado de fallecimiento}
\item{ciudad_fallecimiento}{Ciudad de fallecimiento}
\item{nombre}{Nombre}
\item{apellido}{Apellido}
\item{nombre_completo}{Nombre y apellido}
\item{peso}{Peso en libras}
\item{estatura}{Estatura en pulgadas}
\item{bateador}{Si como bateador es diestro (D) o zurdo (Z)}
\item{lanzador}{Si como lanzador es diestro (D) o zurdo (Z)}
\item{fecha_primer_juego}{Fecha primer juego en primera división}
\item{fecha_ultimo_juego}{Fecha último juego en primera división}
\item{id_retrosheet}{Código de jugador usado en Retrosheet (http://www.retrosheet.org/)}
\item{id_brw}{Código de jugador usado en Baseball Reference website (http://www.baseball-reference.com/)}
\item{fecha_fallecimiento}{Fecha de fallecimiento en formato yyyy-mm-dd}
\item{fecha_nacimiento}{Fecha de nacimiento en formato yyyy-mm-dd}
}}
\description{Nombres de jugadores, fecha de nacimiento e información bibliográfica. Este archivo se usa para obtener detalles acerca de los jugadores listados en bateadores, lanzadores y otros archivos en que los jugadores están identificados por id_jugador.}
\keyword{datasets}
