\docType{data}
\name{pinguinos}
\alias{pinguinos}
\title{Medidas de tamaño de pingüinos adultos en busca de comida cerca de la estación Palmer en la Antártica.}
\format{Un tibble con 344 filas y 8 variables
\describe{
\item{especie}{un factor que indica la especie de pingüino (Adelia, Barbijo y Papúa)}
\item{isla}{un factor que indica la isla en el archipiélago de Palmer (Biscoe, Dream o Torgersen)}
\item{largo_pico_mm}{un número que indica el largo del pico (milímetros)}
\item{alto_pico_mm}{un número que indica el alto del pico (milímetros)}
\item{largo_aleta_mm}{un entero que indica el largo de la aleta (milímetros)}
\item{masa_corporal_g}{un entero que indica la masa corporal (gramos)}
\item{sexo}{un factor que indica el sexo del pingüino (hembra, macho)}
\item{anio}{un entero que indica el año estudiado (2007, 2008 o 2009)}
}}
\usage{pinguinos}
\description{Incluye medidas para: especie de pingüino, isla dentro del Archipiélago de Palmer, tamaño (largo de aleta, masa corporal, dimensiones de pico) y sexo.}
\source{ *** Pingüinos Adelia: Palmer Station Antarctica LTER and K. Gorman. 2020. Structural size measurements and isotopic signatures of foraging among adult male and female Adélie penguins (Pygoscelis adeliae) nesting along the Palmer Archipelago near Palmer Station, 2007-2009 ver 5. Environmental Data Initiative https://doi.org/10.6073/pasta/98b16d7d563f265cb52372c8ca99e60f. *** Pingüinos Papúa: Palmer Station Antarctica LTER and K. Gorman. 2020. Structural size measurements and isotopic signatures of foraging among adult male and female Gentoo penguin (Pygoscelis papua) nesting along the Palmer Archipelago near Palmer Station, 2007-2009 ver 5. Environmental Data Initiative https://doi.org/10.6073/pasta/7fca67fb28d56ee2ffa3d9370ebda689. *** Pingüínos Barbijo: Palmer Station Antarctica LTER and K. Gorman. 2020. Structural size measurements and isotopic signatures of foraging among adult male and female Chinstrap penguin (Pygoscelis antarcticus) nesting along the Palmer Archipelago near Palmer Station, 2007-2009 ver 6. Environmental Data Initiative https://doi.org/10.6073/pasta/c14dfcfada8ea13a17536e73eb6fbe9e. *** Publicado originalmente en: Gorman KB, Williams TD, Fraser WR (2014) Ecological Sexual Dimorphism and Environmental Variability within a Community of Antarctic Penguins (Genus Pygoscelis). PLoS ONE 9(3): e90081. doi:10.1371/journal.pone.0090081}
\keyword{datasets}
