\docType{data}
\name{premios_capitanes}
\alias{premios_capitanes}
\title{Tabla de premios de los capitanes}
\format{Un data frame con 179 filas y 6 columnas
\describe{
\item{id_jugador}{ID del jugador}
\item{nombre_premio}{Nombre del premio recibido}
\item{anio}{Año}
\item{id_liga}{ID liga (factor)}
\item{empate}{Tiene valor "S" (sí) para los premios que corresponden a un empate y "N" (no) en caso contrario.}
\item{notas}{Comentarios sobre el premio}
}}
\description{Información de los premios de los capitanes de equipo}
\keyword{datasets}
