\docType{data}
\name{premios_dirigentes}
\alias{premios_dirigentes}
\title{Tabla de premios de los dirigentes}
\format{Un data frame con 179 filas y 6 columnas
\describe{
\item{id_dirigente}{ID del dirigente}
\item{nombre_premio}{Nombre del premio recibido por el dirigente}
\item{anio}{Año}
\item{id_liga}{ID liga (factor)}
\item{empate}{Tiene valor "S" (sí) para los premios que corresponden a un empate y "N" (no) en caso contrario.}
\item{notas}{Comentarios sobre el premio}
}}
\description{Información de los premios de los dirigentes de equipo}
\keyword{datasets}
