\docType{data}
\name{presidencial}
\alias{presidencial}
\title{Periodos de 11 presidentes, desde Eisenhower a Obama}
\format{Un data frame con 11 filas y 4 variables
\describe{
\item{nombre}{Nombre del presidente}
\item{inicio}{Fecha de inicio de su periodo presidencial}
\item{fin}{Fecha de termino de su periodo presidencial}
\item{partido}{Partido politico al que pertenece}
}}
\description{Nombre, fecha de inicio y termino de su mandato y partido politico de 11 presidentes de EE.UU., desde Eisenhower a Obama.}
\keyword{datasets}
