\docType{data}
\name{representantes}
\alias{representantes}
\title{Tabla de representantes (bateadores)}
\format{Un data frame con 3.504 filas y 10 columnas
\describe{
\item{id_jugador}{ID del jugador}
\item{anio}{Año}
\item{id_equipo}{ID equipo (factor)}
\item{id_liga}{ID liga (factor)}
\item{en_temporada}{Cero si representó al equipo toda la temporada. En otro caso denota el orden de entrada en la temporada (uno si fue el primer representante que entró esa temporada, dos si fue el segundo representante que entró esa temporada, etc)}
\item{partidos}{Partidos dirigidos}
\item{partidos_ganados}{Partidos ganados}
\item{partidos_perdidos}{Partidos perdidos}
\item{posicion}{Posición del equipo en la clasificación final del año}
\item{jugador_representado}{Tiene valor "S" (sí) para los jugadores que tienen representante en la temporada y "N" (no) en caso contrario (factor).}
}}
\description{Información de los representantes de los bateadores}
\keyword{datasets}
