\docType{data}
\name{salarios}
\alias{salarios}
\title{Tabla de salarios}
\format{Un data frame con 26.428 filas y 5 columnas
\describe{
\item{anio}{Año}
\item{id_equipo}{ID equipo (factor)}
\item{id_liga}{ID liga (factor)}
\item{id_jugador}{ID del jugador}
\item{salario}{Salario en dólares}
}}
\description{Información de los salarios de los jugadores.}
\keyword{datasets}
