\docType{data}
\name{sentences}
\alias{sentences}
\title{Muestra de vector de caracteres para practicar manipulacion de cadena de caracteres.}
\format{Una cadena de caracteres con 720 elmentos.
\describe{
}}
\usage{sentences}
\description{`sentences` es una colección de "Harvard sentences" usada para pruebas estandarizadas de voz.}
\keyword{datasets}
