\docType{data}
\name{tabla1}
\alias{tabla1}
\title{Registros de tuberculosis de la Organizacion Mundial de la salud (1era variante)}
\format{Un tibble con 6 filas y 4 columnas
\describe{
\item{pais}{Nombre del pais}
\item{anio}{Ano correspondiente a los valores registrados}
\item{casos}{Numero de casos registrados}
\item{poblacion}{Poblacion del pais}
}}
\description{Contiene observaciones de casos registrados en Afganistan, Brasil y China en los anos 1999 y 2000.}
\keyword{datasets}
