\docType{data}
\name{tabla2}
\alias{tabla2}
\title{Registros de tuberculosis de la Organizacion Mundial de la salud (2da variante)}
\format{Un tibble con 12 filas y 4 columnas
\describe{
\item{pais}{Nombre del pais}
\item{anio}{Ano correspondiente a los valores registrados}
\item{tipo}{Categoria a la que pertenece la observacion}
\item{cuenta}{Cantidad de observaciones para cada tipo}
}}
\description{Contiene observaciones de casos registrados en Afganistan, Brasil y China en los anos 1999 y 2000.}
\keyword{datasets}
