\docType{data}
\name{tabla4a}
\alias{tabla4a}
\title{Registros de tuberculosis de la Organizacion Mundial de la salud (3era variante)}
\format{Un tibble con 3 filas y 3 columnas
\describe{
\item{pais}{Nombre del pais}
\item{1999}{Casos registrados el ano 1999.}
\item{2000}{Casos registrados el ano 2000.}
}}
\description{Contiene observaciones de casos registrados en Afganistan, Brasil y China en los anos 1999 y 2000.}
\keyword{datasets}
