\docType{data}
\name{tabla5}
\alias{tabla5}
\title{Registros de tuberculosis de la Organizacion Mundial de la salud (3era variante)}
\format{Un tibble con 6 filas y 4 columnas
\describe{
\item{pais}{Nombre del pais}
\item{siglo}{Siglo correspondiente a los valores registrados}
\item{anio}{Ano correspondiente a los valores registrados}
\item{tasa}{Tasa de casos de tuberculosis en relacion a la poblacion total}
}}
\description{Contiene observaciones de casos registrados en Afganistan, Brasil y China en los anos 1999 y 2000.}
\keyword{datasets}
