% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{translate}
\alias{translate}
\title{Avoid R CMD check warning}
\format{
An object of class \code{function} of length 1.
}
\usage{
translate
}
\description{
Avoid R CMD check warning
}
\keyword{internal}
