\docType{data}
\name{vehiculos}
\alias{vehiculos}
\title{Datos de economía de combustible}
\format{Un data.frame con 33442 filas y 12 columnas
\describe{
\item{id}{Identificador único EPA}
\item{fabricante}{Fabricante}
\item{modelo}{Nombre del modelo}
\item{anio}{año de fabricación}
\item{clase}{tipo de auto}
\item{transmision}{tipo de transmisión}
\item{traccion}{tipo de tracción}
\item{cilindros}{número de cilindros}
\item{motor}{tamaño del motor del automóvil, en litros}
\item{combustible}{tipo de combustible}
\item{autopista}{millas por galón de combustible en autopista}
\item{ciudad}{millas por galón de combustible en ciudad}
}}
\usage{vehiculos}
\description{Datos de economía de combustible de la Agencia de Protección Medioambiental (EPA) de EE.UU., 1985-2015. Contiene una selección de variables y no considera vehículos con datos incompletos.}
\keyword{datasets}
