\docType{data}
\name{aerolineas}
\alias{aerolineas}
\title{Nombres de aerolíneas}
\format{Un data.frame con 16 filas y 2 columnas
\describe{
\item{aerolinea}{abreviación de dos caracteres del nombre de la aerolínea}
\item{nombre}{nombre completo de la aerolínea}
}}
\usage{aerolineas}
\description{Nombres de aerolíneas y su respectivo código carrier de dos dígitos.}
\keyword{datasets}
