\docType{data}
\name{aeropuertos}
\alias{aeropuertos}
\title{Datos de aeropuertos}
\format{Un data.frame con 1.458 filas y 8 columnas
\describe{
\item{codigo_aeropuerto}{código FAA del aeropuerto}
\item{nombre}{nombre del aeropuerto}
\item{latitud}{coordenada de ubicación del aeropuerto}
\item{longitud}{coordenada de ubicación del aeropuerto}
\item{altura}{altura medida en pies}
\item{zona_horaria}{diferencia horaria respecto del meridiano de Greenwich (GMT)}
\item{horario_verano}{horario legal de acuerdo a las políticas de ahorro de energía. A = Zona Horaria de Ahorro de Luz de Día (DST) estándar de los EEUU: comienza el segundo Sábado de Marzo y termina el primer Sábado de Noviembre. U = Desconocido. N = No tiene zona horaria especial.}
\item{zona_horaria_iana}{zona horaria IANA, la cual se determina de acuerdo al servicio web GeoNames}
}}
\usage{aeropuertos}
\description{Metadatos útiles de los aeropuertos}
\keyword{datasets}
