\docType{data}
\name{aviones}
\alias{aviones}
\title{Datos de aviones}
\format{Un data.frame con 3.322 filas y 9 columnas
\describe{
\item{codigo_cola}{código de cola}
\item{anio}{año de fabricación}
\item{tipo}{tipo de avión}
\item{fabricante}{fabricante del avión}
\item{modelo}{modelo del avión}
\item{motores}{cantidad de motores}
\item{asientos}{cantidad de asientos}
\item{velocidad}{velocidad crucero promedio medida en millas por hora}
\item{tipo_motor}{tipo de motor del avión}
}}
\usage{aviones}
\description{Metadatos de los aviones y sus códigos de cola en el registro de aviación de la FAA. American Airways (AA) y Envoy Air (MQ) reportan número de flota en lugar de número de cola, por lo que no es posible trazarlos.}
\keyword{datasets}
