\docType{data}
\name{bateadores}
\alias{bateadores}
\title{Tabla de estadísticas de bateo}
\format{Un data frame con 105.861 filas y 22 columnas
\describe{
\item{id_jugador}{ID del jugador}
\item{id_anio}{Año}
\item{orden_equipos}{Orden en que el jugador se movió entre equipos dentro de la misma temporada}
\item{id_equipo}{ID del equipo (factor)}
\item{id_liga}{ID de la liga (factor con niveles AA, AL, FL, NL, PL, UA)}
\item{juegos}{Número de juegos en los que participó oficialmente el jugador}
\item{al_bate}{Turnos al bate}
\item{carreras}{Carreras anotadas}
\item{golpes}{Cantidad de veces en la que el jugador alcanzó una base debido a un golpe, sin error de la defensa}
\item{dobles}{Hits en los que el bateador llegó a segunda base}
\item{triples}{Golpes en los que el bateador llegó a tercera base}
\item{cuadrangulares}{Cuadrangulares - jonrones}
\item{carreras_empujadas}{Carreras empujadas}
\item{bases_robadas}{Bases robadas}
\item{atrapado_robando}{Atrapado (out) en intento de robo}
\item{base_bolas}{Base por bolas}
\item{ponches}{Ponches}
\item{base_intencional}{Base por bolas intencional}
\item{golpeado}{Golpeado por lanzamiento}
\item{toque_sacrificio}{Toque de sacrificio}
\item{elavado_sacrificio}{Elevado de sacrificio}
\item{doble_matanza}{Batazos para doble matanza (doble play)}
}}
\description{Estadísticas de bateadores}
\keyword{datasets}
