\docType{data}
\name{capitanes}
\alias{capitanes}
\title{Tabla de capitanes}
\format{Un data frame con 3.504 filas y 10 columnas
\describe{
\item{id_jugador}{ID del jugador}
\item{anio}{Año}
\item{id_equipo}{ID equipo (factor)}
\item{id_liga}{ID liga (factor)}
\item{en_temporada}{Cero si fue capitán del equipo al equipo toda la temporada. En otro caso denota el orden de entrada en la temporada (uno si fue el primer capitán del equipo que entró esa temporada, dos si fue el segundo capitán del equipo que entró esa temporada, etc)}
\item{juegos}{Juegos dirigidos}
\item{juegos_ganados}{Juegos ganados}
\item{juegos_perdidos}{Juegos perdidos}
\item{posicion}{Posición del equipo en la clasificación final del año}
\item{jugador_representado}{Tiene valor "S" (sí) para los jugadores estuvieron de capitán del equipo durante la temporada y "N" (no) en caso contrario (factor).}
}}
\description{Información de los equipos que dirigieron y algunas estadísticas básicas de los equipos en cada año.}
\keyword{datasets}
