\docType{data}
\name{clima}
\alias{clima}
\title{Datos de clima}
\format{Un data.frame con 3,322 líneas y 9 variables
\describe{
\item{origen}{Estación de clima}
\item{anio}{año de la fecha de registro}
\item{mes}{mes de la fecha de registro}
\item{dia}{día de la fecha de registro}
\item{hora}{hora de la fecha de registro}
\item{temperatura}{temperatura medida en grados Fahrenheit}
\item{punto_rocio}{temperatura medida en grados Fahrenheit en la que, dada la presión atmosférica y humedad, condensa el vapor de agua}
\item{humedad}{humedad relativa}
\item{direccion_viento}{dirección del viento medida en grados}
\item{velocidad_viento}{velocidad del viento medida en millas por hora}
\item{velocidad_rafaga}{velocidad de las ráfagas de viento medida en millas por hora}
\item{precipitacion}{precipitación de lluvia medida en pulgadas}
\item{presion}{presión atmosférica al nivel del mar medida en milibares}
\item{visibilidad}{visibilidad medida en millas}
\item{fecha_hora}{fecha y hora de registro en formato POSIXct}
}}
\usage{clima}
\description{Datos meteorológicos por hora para LGA, JFK y EWR.}
\keyword{datasets}
