\docType{data}
\name{comunes}
\alias{comunes}
\title{Modelos comunes de vehículos}
\format{Un data.frame con 347 filas y 4 columnas
\describe{
\item{fabricante}{Fabricante}
\item{modelo}{Modelo}
\item{total_modelos}{Número total de modelos}
\item{total_modelos_anio}{Número total de modelos-año}
}}
\usage{comunes}
\description{Estos modelos tienen al menos 10 años de datos. Este dataset es adecuado para efectuar un join que restringa el número de observaciones a un nivel manejable.}
\keyword{datasets}
