% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datos-package.R
\docType{package}
\name{datos-package}
\alias{datos}
\alias{datos-package}
\title{datos: Translates various data sets into Spanish}
\description{
It translates the name of the data set, its variables, and values. 
  It focuses on the practice data sets used in the book 'R for Data Science'. 
  It includes translations into Spanish for: 'airlines', 'airports', 'Batting', 
  'diamonds', 'faithful', 'flights', 'gapminder', 'gss_cat', 'mpg', 'mtcars', 
  'planes', 'presidential', 'table1', 'table2', 'table3', 'table4a', 'table4b', 
  'table5', 'weather', 'who'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cienciadedatos/datos}
  \item Report bugs at \url{https://github.com/cienciadedatos/datos/issues}
}

}
\author{
\strong{Maintainer}: Edgar Ruiz \email{edgararuiz@gmail.com}

Authors:
\itemize{
  \item Riva Quiroga \email{riva.quiroga@uc.cl}
  \item Mauricio Vargas \email{mvargas@dcc.uchile.cl}
  \item Mauro Lepore \email{maurolepore@gmail.com}
}

}
\keyword{internal}
