% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datos.R
\docType{package}
\name{datos-package}
\alias{datos}
\alias{datos-package}
\title{datos: Traduce al Español Varios Conjuntos de Datos de Práctica}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

Provee una versión traducida de los siguientes conjuntos de datos: 
  'airlines', 'airports', 'babynames', 'Batting', 'diamonds', 'faithful', 'flights', 'gapminder', 
  'gss_cat', 'iris', 'mpg', 'mtcars', 'atmos', 'planes', 'presidential', 'table1', 'table2', 'table3', 
  'table4a', 'table4b', 'table5', 'vehicles','weather', 'who'. English: It provides a
  Spanish translated version of the datasets listed above.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cienciadedatos/datos}
  \item Report bugs at \url{https://github.com/cienciadedatos/datos/issues}
}

}
\author{
\strong{Maintainer}: Edgar Ruiz \email{edgararuiz@gmail.com}

Authors:
\itemize{
  \item Riva Quiroga \email{riva.quiroga@uc.cl}
  \item Mauricio Vargas \email{mvargas@dcc.uchile.cl}
  \item Mauro Lepore \email{maurolepore@gmail.com}
}

Other contributors:
\itemize{
  \item Rayna Harris \email{rayna.harris@gmail.com} [contributor]
  \item Daniela Vasquez \email{daniela.vazquez@gmail.com} [contributor]
}

}
\keyword{internal}
