% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datos.R
\docType{package}
\name{datos-package}
\alias{datos}
\alias{datos-package}
\title{datos: Traduce al Español Varios Conjuntos de Datos de Práctica}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Provee una versión traducida de los siguientes conjuntos de datos: 'airlines', 'airports', 'AwardsManagers', 'babynames', 'Batting', 'diamonds', 'faithful', 'fueleconomy', 'Fielding', 'flights', 'gapminder', 'gss_cat', 'iris', 'Managers', 'mpg', 'mtcars', 'atmos', 'palmerpenguins', 'People, 'Pitching', 'planes', 'presidential', 'table1', 'table2', 'table3', 'table4a', 'table4b', 'table5', 'vehicles', 'weather', 'who'. English: It provides a Spanish translated version of the datasets listed above.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cienciadedatos/datos}
  \item Report bugs at \url{https://github.com/cienciadedatos/datos/issues}
}

}
\author{
\strong{Maintainer}: Riva Quiroga \email{riva.quiroga@uc.cl} (\href{https://orcid.org/0000-0002-1147-4135}{ORCID})

Authors:
\itemize{
  \item Edgar Ruiz \email{edgararuiz@gmail.com}
  \item Mauricio Vargas \email{mvargas@dcc.uchile.cl}
  \item Mauro Lepore \email{maurolepore@gmail.com} (\href{https://orcid.org/0000-0002-1986-7988}{ORCID})
}

Other contributors:
\itemize{
  \item Rayna Harris \email{rayna.harris@gmail.com} [contributor]
  \item Daniela Vasquez \email{daniela.vazquez@gmail.com} [contributor]
}

}
\keyword{internal}
