\docType{data}
\name{diamantes}
\alias{diamantes}
\title{Precio de 50,000 diamantes}
\format{Un data.frame con 53,940 lineas y 10 variables
\describe{
\item{precio}{Precio en dolares US}
\item{quilate}{Peso del diamante}
\item{corte}{Calided del corte}
\item{claridad}{Medida de que tan claro es el diamante}
\item{profundidad}{Porcentaje total de la profundidad}
\item{tabla}{Medida de la parte mas ancha del diamante}
\item{x}{Largo in milimetros}
\item{y}{Ancho in milimetros}
\item{z}{Profundidad en milimetros}
}}
\usage{diamantes}
\description{Un set que contiene los precios de casi 54,000 diamantes.}
\keyword{datasets}
