\docType{data}
\name{diamantes}
\alias{diamantes}
\title{Precio de 50.000 diamantes}
\format{Un data.frame con 53.940 filas y 10 columnas
\describe{
\item{precio}{Precio en dólares americanos (\$326-\$18,823)}
\item{quilate}{Peso del diamante (0.2-5.01)}
\item{corte}{Calidad del corte (Regular, Bueno, Muy bueno, Premium, Ideal)}
\item{color}{Color del diamante, de J (peor) a D (mejor)}
\item{claridad}{Medida de que tan claro es el diamante (I1 (peor), SI1, SI2, VS1, VS2, VVS1, VVS2, IF (mejor))}
\item{profundidad}{Porcentaje de la profundidad total = z / mean(x, y) = 2 * z / (x + y) (43-79)}
\item{tabla}{Ancho de la parte superior del diamante con relación a su punto más ancho (43-95)}
\item{x}{Largo en milimetros}
\item{y}{Ancho en milimetros}
\item{z}{Profundidad en milimetros}
}}
\usage{diamantes}
\description{Un conjunto de datos que contiene los precios de casi 54.000 diamantes.}
\keyword{datasets}
