\docType{data}
\name{encuesta}
\alias{encuesta}
\title{Muestra de variables categóricas de una encuesta social}
\format{Un data.frame con 21.473 filas y 9 columnas
\describe{
\item{anio}{año de la encuesta, 2000-2014}
\item{estado_civil}{estado civil}
\item{edad}{edad. Edad máxima truncada a 89 años}
\item{raza}{categorías raciales según la encuesta}
\item{ingreso}{ingreso reportado por los encuestados (US$)}
\item{partido}{partido político con el que se identifica}
\item{religion}{religión}
\item{denominacion}{denominación religiosa a la que pertenece}
\item{horas_tv}{horas de televisiín que ve al día}
}}
\usage{encuesta}
\description{Muestra de variables categóricas de la Encuesta Social General de EE.UU.}
\keyword{datasets}
