\docType{data}
\name{fiel}
\alias{fiel}
\title{Datos del géiser Viejo Fiel (Old Faithful)}
\format{Un data frame con 272 observaciones y 2 columnas
\describe{
\item{erupciones}{Tiempo de erupciones, en minutos}
\item{espera}{Minutos entre erupciones}
}}
\usage{fiel}
\description{Datos de tiempo de duración y espera entre erupciones del géiser Viejo Fiel (Old Faithful), ubicado en el Parque Nacional de Yellowstone, en Wyoming, Estados Unidos.}
\keyword{datasets}
