\docType{data}
\name{lanzadores}
\alias{lanzadores}
\title{Tabla de estadísticas de lanzadores}
\format{Un data frame con 46.699 filas y 30 columnas
\describe{
\item{id_jugador}{ID del jugador}
\item{id_anio}{Año}
\item{orden_equipos}{Orden en que el jugador se movió entre equipos dentro de la misma temporada}
\item{id_equipo}{ID del equipo (factor)}
\item{id_liga}{ID de la liga (factor con niveles AA, AL, FL, NL, PL, UA)}
\item{juegos_ganados}{Juegos jugados ganados}
\item{juegos_perdidos}{Juegos jugados perdidos}
\item{juegos_jugados}{Juegos jugados}
\item{juegos_iniciados}{Juegos jugados iniciados}
\item{juegos_completados}{Número de juegos completos (9 innings lanzados) que lanzó el jugador}
\item{blanqueadas}{Número de blanqueos (juegos completos sin permitir carreras) que lanzó el jugador}
\item{juegos_salvados}{Juegos salvados}
\item{IPouts}{Outs propinados al equipo contrario (Igual a innings lanzados x 3)}
\item{hits}{Hits permitidos del oponente}
\item{carreras_ganadas}{Carreras limpias recibidas}
\item{cuadrangulares}{Cuadrangulares recibidos}
\item{BB}{Base por bolas cedidas al oponente}
\item{ponches}{Ponches propinados al oponente}
\item{promedio_bateo_rival}{Promedio de bateo del rival}
\item{promedio_carreras_ganadas}{Promedio de carreras limpias permitidas (normalizada a 9 innings lanzados)}
\item{IBB}{Base por bolas intencionales cedidas al oponente}
\item{lanzamientos_desviados}{Lanzamientos desviados lanzados por el lanzador}
\item{HBP}{Bateadores golpeador por el lanzador}
\item{BK}{Balks (movimiento ilegal del cuerpo realizado por el lanzador)}
\item{BFP}{Bateadores a los que el lanzador se ha enfrentado}
\item{juegos_finalizados}{Juegos en los que el lanzador finalizó el juego}
\item{carreras}{Carreras recebidas (sucias y limpias)}
\item{sacrificios_golpeados}{Toques de sacrificio que el oponente le hizo al lanzador}
\item{vuelos_sacrificio}{Elevado (fly) de sacrificio que el oponente le hizo al lanzador}
\item{doble_matanza}{Doble matanza inducidas por el lanzador}
}}
\description{Estadísticas de lanzadores}
\keyword{datasets}
