\docType{data}
\name{millas}
\alias{millas}
\title{Datos de economía de combustible de 1999 y 2008 para 38 modelos populares de automóviles}
\format{Un data.frame con 234 filas y 11 columnas
\describe{
\item{fabricante}{fabricante}
\item{modelo}{nombre del modelo}
\item{cilindrada}{tamaño del motor del automóvil, en litros}
\item{anio}{año de fabricación}
\item{cilindros}{número de cilindros}
\item{transmision}{tipo de transmisión}
\item{traccion}{tipo de tracción (d = delantera, t = trasera, 4 = 4 ruedas)}
\item{ciudad}{millas por galón de combustible en ciudad}
\item{autopista}{millas por galón de combustible en autopista}
\item{combustible}{tipo de combustible (p = premium, r = regular, e = etanol, d = diesel, g = gas natural comprimido)}
\item{clase}{tipo de auto}
}}
\usage{millas}
\description{Este conjunto de datos contiene un subconjunto de los datos de economía de combustible que la Agencia de Protección Medioambiental (EPA) pone a disposición en http://fueleconomy.gov. Contiene solo modelos que tuvieron una nueva versión cada año entre 1999 y 2008, lo que fue utilizado como un proxy de la popularidad del modelo.}
\keyword{datasets}
