\docType{data}
\name{mtautos}
\alias{mtautos}
\title{Pruebas de ruta de automóviles de Motor Trend}
\format{Un data.frame con 32 filas y 12 columnas
\describe{
\item{modelo}{modelo (y fabricante) del vehículo}
\item{millas}{millas por galón de Estados Unidos}
\item{cilindros}{número de cilindros}
\item{cilindrada}{suma del volumen útil de todos los cilindros del motor en pulgadas cúbicas}
\item{caballos}{caballos de fuerza brutos}
\item{eje}{relación del eje de transmisión trasero}
\item{peso}{peso (1000 libras)}
\item{velocidad}{tiempo en recorrer 1/4 de milla}
\item{forma}{forma del motor (en V o en línea)}
\item{transmision}{tipo de transmisión (0 = automático, 1 = manual)}
\item{cambios}{número de cambios de la caja de cambios}
\item{carburadores}{número de carburadores}
}}
\usage{mtautos}
\description{Los datos fueron extraídos de la revista Motor Trend de Estados Unidos de 1974, y tiene datos de consumo de combustible y 10 aspectos de diseño y rendimiento de automóviles para 32 automóviles (modelos de 1973-1974).}
\keyword{datasets}
