\docType{data}
\name{oms}
\alias{oms}
\title{Datos de tuberculosis de la Organización Mundial de la Salud}
\format{Un data.frame con 7.240 filas y 60 columnas
\describe{
\item{pais}{nombre del país}
\item{iso2}{código ISO de dos dígitos del país}
\item{iso3}{código ISO de tres dígitos del país}
\item{anio}{año}
\item{nuevos_fpp_h014}{casos nuevos, diagnóstico frotis pulmonar positivo (fpp), hombres (h) de 0 a 14 años (014)}
\item{nuevos_fpp_h1524}{casos nuevos, diagnóstico frotis pulmonar positivo (fpp), hombres (h) de 15 a 24 años (1524)}
\item{nuevos_fpp_h2534}{casos nuevos, diagnóstico frotis pulmonar positivo (fpp), hombres (h) de 25 a 34 años (2534)}
\item{nuevos_fpp_h3544}{casos nuevos, diagnóstico frotis pulmonar positivo (fpp), hombres (h) de 35 a 44 años (3544)}
\item{nuevos_fpp_h4554}{casos nuevos, diagnóstico frotis pulmonar positivo (fpp), hombres (h) de 45 a 54 años (4554)}
\item{nuevos_fpp_h5564}{casos nuevos, diagnóstico frotis pulmonar positivo (fpp), hombres (h) de 55 a 64 años (5564)}
\item{nuevos_fpp_h65}{casos nuevos, diagnóstico frotis pulmonar positivo (fpp), hombres (h) de 65 o más años}
\item{nuevos_fpp_m014}{casos nuevos, diagnóstico frotis pulmonar positivo (fpp), mujeres (m) de 0 a 14 años (014)}
\item{nuevos_fpp_m1524}{casos nuevos, diagnóstico frotis pulmonar positivo (fpp), mujeres (m) de 15 a 24 años (1524)}
\item{nuevos_fpp_m2534}{casos nuevos, diagnóstico frotis pulmonar positivo (fpp), mujeres (m) de 25 a 34 años (2534)}
\item{nuevos_fpp_m3544}{casos nuevos, diagnóstico frotis pulmonar positivo (fpp), mujeres (m) de 35 a 44 años (3544)}
\item{nuevos_fpp_m4554}{casos nuevos, diagnóstico frotis pulmonar positivo (fpp), mujeres (m) de 45 a 54 años (4554)}
\item{nuevos_fpp_m5564}{casos nuevos, diagnóstico frotis pulmonar positivo (fpp), mujeres (m) de 55 a 64 años (5564)}
\item{nuevos_fpp_m65}{casos nuevos, diagnóstico frotis pulmonar positivo (fpp), mujeres (m) de 65 o más años}
\item{nuevos_fpn_h014}{casos nuevos, diagnóstico frotis pulmonar negativo (fpn), hombres (h) de 0 a 14 años (014)}
\item{nuevos_fpn_h1524}{casos nuevos, diagnóstico frotis pulmonar negativo (fpn), hombres (h) de 15 a 24 años (1524)}
\item{nuevos_fpn_h2534}{casos nuevos, diagnóstico frotis pulmonar negativo (fpn), hombres (h) de 25 a 34 años (2534)}
\item{nuevos_fpn_h3544}{casos nuevos, diagnóstico frotis pulmonar negativo (fpn), hombres (h) de 35 a 44 años (3544)}
\item{nuevos_fpn_h4554}{casos nuevos, diagnóstico frotis pulmonar negativo (fpn), hombres (h) de 45 a 54 años (4554)}
\item{nuevos_fpn_h5564}{casos nuevos, diagnóstico frotis pulmonar negativo (fpn), hombres (h) de 55 a 64 años (5564)}
\item{nuevos_fpn_h65}{casos nuevos, diagnóstico frotis pulmonar negativo (fpn), hombres (h) de 65 o más años}
\item{nuevos_fpn_m014}{casos nuevos, diagnóstico frotis pulmonar negativo (fpn), mujeres (m) de 0 a 14 años (014)}
\item{nuevos_fpn_m1524}{casos nuevos, diagnóstico frotis pulmonar negativo (fpn), mujeres (m) de 15 a 24 años (1524)}
\item{nuevos_fpn_m2534}{casos nuevos, diagnóstico frotis pulmonar negativo (fpn), mujeres (m) de 25 a 34 años (2534)}
\item{nuevos_fpn_m3544}{casos nuevos, diagnóstico frotis pulmonar negativo (fpn), mujeres (m) de 35 a 44 años (3544)}
\item{nuevos_fpn_m4554}{casos nuevos, diagnóstico frotis pulmonar negativo (fpn), mujeres (m) de 45 a 54 años (4554)}
\item{nuevos_fpn_m5564}{casos nuevos, diagnóstico frotis pulmonar negativo (fpn), mujeres (m) de 55 a 64 años (5564)}
\item{nuevos_fpn_m65}{casos nuevos, diagnóstico frotis pulmonar negativo (fpn), mujeres (m) de 65 o más años}
\item{nuevos_ep_h014}{casos nuevos, diagnóstico extrapulmonar (ep), hombres (h) de 0 a 14 años (014)}
\item{nuevos_ep_h1524}{casos nuevos, diagnóstico extrapulmonar (ep), hombres (h) de 15 a 24 años (1524)}
\item{nuevos_ep_h2534}{casos nuevos, diagnóstico extrapulmonar (ep), hombres (h) de 25 a 34 años (2534)}
\item{nuevos_ep_h3544}{casos nuevos, diagnóstico extrapulmonar (ep), hombres (h) de 35 a 44 años (3544)}
\item{nuevos_ep_h4554}{casos nuevos, diagnóstico extrapulmonar (ep), hombres (h) de 45 a 54 años (4554)}
\item{nuevos_ep_h5564}{casos nuevos, diagnóstico extrapulmonar (ep), hombres (h) de 55 a 64 años (5564)}
\item{nuevos_ep_h65}{casos nuevos, diagnóstico extrapulmonar (ep), hombres (h) de 65 o más años}
\item{nuevos_ep_m014}{casos nuevos, diagnóstico extrapulmonar (ep), mujeres (m) de 0 a 14 años (014)}
\item{nuevos_ep_m1524}{casos nuevos, diagnóstico extrapulmonar (ep), mujeres (m) de 15 a 24 años (1524)}
\item{nuevos_ep_m2534}{casos nuevos, diagnóstico extrapulmonar (ep), mujeres (m) de 25 a 34 años (2534)}
\item{nuevos_ep_m3544}{casos nuevos, diagnóstico extrapulmonar (ep), mujeres (m) de 35 a 44 años (3544)}
\item{nuevos_ep_m4554}{casos nuevos, diagnóstico extrapulmonar (ep), mujeres (m) de 45 a 54 años (4554)}
\item{nuevos_ep_m5564}{casos nuevos, diagnóstico extrapulmonar (ep), mujeres (m) de 55 a 64 años (5564)}
\item{nuevos_ep_m65}{casos nuevos, diagnóstico extrapulmonar (ep), mujeres (m) de 65 o más años}
\item{nuevosrecaida_h014}{nuevas recaídas, hombres (h) de 0 a 14 años (014)}
\item{nuevosrecaida_h1524}{nuevas recaídas, hombres (h) de 15 a 24 años (1524)}
\item{nuevosrecaida_h2534}{nuevas recaídas, hombres (h) de 25 a 34 años (2534)}
\item{nuevosrecaida_h3544}{nuevas recaídas, hombres (h) de 35 a 44 años (3544)}
\item{nuevosrecaida_h4554}{nuevas recaídas, hombres (h) de 45 a 54 años (4554)}
\item{nuevosrecaida_h5564}{nuevas recaídas, hombres (h) de 55 a 64 años (5564)}
\item{nuevosrecaida_h65}{nuevas recaídas, hombres (h) de 65 o más años}
\item{nuevosrecaida_m014}{nuevas recaídas, mujeres (m) de 0 a 14 años (014)}
\item{nuevosrecaida_m1524}{nuevas recaídas, mujeres (m) de 15 a 24 años (1524)}
\item{nuevosrecaida_m2534}{nuevas recaídas, mujeres (m) de 25 a 34 años (2534)}
\item{nuevosrecaida_m3544}{nuevas recaídas, mujeres (m) de 35 a 44 años (3544)}
\item{nuevosrecaida_m4554}{nuevas recaídas, mujeres (m) de 45 a 54 años (4554)}
\item{nuevosrecaida_m5564}{nuevas recaídas, mujeres (m) de 55 a 64 años (5564)}
\item{nuevosrecaida_m65}{nuevas recaídas, mujeres (m) de 65 o más años}
}}
\usage{oms}
\description{Subconjunto de datos del informe anual de tuberculosis de la Organización Mundial de la Salud}
\keyword{datasets}
