\docType{data}
\name{oraciones}
\alias{oraciones}
\title{Muestra de vectores de caracteres para practicar manipulación de cadenas}
\format{Un tibble con 57 filas y 1 columna
\describe{
\item{colores}{Oraciones que contienen colores}
}}
\description{Oraciones es una colección de "Harvard sentences" usada en pruebas estandarizadas de voz.}
\keyword{datasets}
