\docType{data}
\name{paises}
\alias{paises}
\title{Datos de Gapminder}
\format{El dataframe principal tiene 1704 filas y 6 variables
\describe{
\item{pais}{países incluidos (factor con 142 niveles)}
\item{continente}{continentes (factor con 5 niveles)}
\item{anio}{desde 1952 a 2007, datos cada 5 años}
\item{esperanza_de_vida}{esperanza de vida al nacer, en años}
\item{poblacion}{población}
\item{pib_per_capita}{PIB per cápita (en dólares US, ajustados según inflación)}
}}
\usage{paises}
\description{Extracto de datos de Gapminder sobre expectativa de vida, PIB per cápita y población, según país}
\keyword{datasets}
