\docType{data}
\name{presidencial}
\alias{presidencial}
\title{Periodos de 12 presidentes, desde Eisenhower a Trump}
\format{Un data.frame con 12 filas y 4 columnas
\describe{
\item{nombre}{Nombre del presidente}
\item{inicio}{Fecha de inicio de su periodo presidencial}
\item{fin}{Fecha de término de su periodo presidencial}
\item{partido}{Partido político al que pertenece}
}}
\description{Nombre, fecha de inicio y término de su mandato y partido político de 11 presidentes de EE.UU., desde Eisenhower a Trump.}
\keyword{datasets}
