\docType{data}
\name{tabla1}
\alias{tabla1}
\title{Registros de tuberculosis de la Organización Mundial de la Salud (1era variante)}
\format{Un tibble con 6 filas y 4 columnas
\describe{
\item{pais}{Nombre del país}
\item{anio}{Año correspondiente a los valores registrados}
\item{casos}{Número de casos registrados}
\item{poblacion}{Población del país}
}}
\description{Contiene observaciones de casos registrados en Afganistán, Brasil y China en los años 1999 y 2000.}
\keyword{datasets}
