\docType{data}
\name{tabla2}
\alias{tabla2}
\title{Registros de tuberculosis de la Organización Mundial de la Salud (2da variante)}
\format{Un tibble con 12 filas y 4 columnas
\describe{
\item{pais}{Nombre del país}
\item{anio}{Año correspondiente a los valores registrados}
\item{tipo}{Categoría a la que pertenece la observación}
\item{cuenta}{Cantidad de observaciones para cada tipo}
}}
\description{Contiene observaciones de casos registrados en Afganistán, Brasil y China en los años 1999 y 2000.}
\keyword{datasets}
