\docType{data}
\name{tabla3}
\alias{tabla3}
\title{Registros de tuberculosis de la Organización Mundial de la Salud (3ra variante)}
\format{Un tibble con 6 filas y 3 columnas
\describe{
\item{pais}{Nombre del país}
\item{anio}{Año correspondiente a los valores registrados}
\item{tasa}{Tasa de casos de tuberculosis en relación a la población total}
}}
\description{Contiene observaciones de casos registrados en Afganistán, Brasil y China en los años 1999 y 2000.}
\keyword{datasets}
