\docType{data}
\name{tabla4a}
\alias{tabla4a}
\title{Registros de tuberculosis de la Organizaci칩n Mundial de la Salud (3era variante)}
\format{Un tibble con 3 filas y 3 columnas
\describe{
\item{pais}{Nombre del pais}
\item{1999}{Casos registrados el a침o 1999.}
\item{2000}{Casos registrados el a침o 2000.}
}}
\description{Contiene observaciones de casos registrados en Afganistan, Brasil y China en los a침os 1999 y 2000.}
\keyword{datasets}
