\docType{data}
\name{tabla4b}
\alias{tabla4b}
\title{Registros de tuberculosis de la Organización Mundial de la Salud (3era variante)}
\format{Un tibble con 3 filas y 3 columnas
\describe{
\item{pais}{Nombre del país}
\item{1999}{Poblacion en el año 1999.}
\item{2000}{Poblacion en el año 2000.}
}}
\description{Contiene observaciones de casos registrados en Afganistán, Brasil y China en los años 1999 y 2000.}
\keyword{datasets}
