\docType{data}
\name{tabla5}
\alias{tabla5}
\title{Registros de tuberculosis de la Organización Mundial de la Salud (5ta variante)}
\format{Un tibble con 6 filas y 4 columnas
\describe{
\item{pais}{Nombre del país}
\item{siglo}{Siglo correspondiente a los valores registrados}
\item{anio}{Año correspondiente a los valores registrados}
\item{tasa}{Tasa de casos de tuberculosis con relación a la población total}
}}
\description{Contiene observaciones de casos registrados en Afganistán, Brasil y China en los años 1999 y 2000.}
\keyword{datasets}
